Reorder fields in track creation form
The track creation form in 'app/views/tracks/new.html.erb' file was re-adjusted. The field for 'track_type' was moved down, falling into a new order of fields. This was done to better organize the form, making it more user-friendly, especially for users not familiar with track information. The changes ensure a smoother process for track creation.