Refactor user profile layout and navigation bar
The layout for user profile in the 'app/views/users/show.html.erb' file was refactored to improve readability and the roles are now filtered to exclude 'admin' roles. Furthermore, the navigation bar was dramatically refactored associated with the implementation of mobile-friendly navigation system. Mobile and desktop views were separated into different files for better organization and responsiveness on different devices sizes. Along with this change, a script was added to toggle between the two menus, depending on the size of the viewing window. This will greatly enhance the user experience on different devices.