Update track condition and clean up unused field
Removed 'time_of_day' and 'weather_condition' fields from app in various places and replaced with 'track_condition' field. This change was made to simplify and streamline track condition information. Also cleaned up some validations, scopes, forms and views to reflect these changes. The views for leaderboard were updated to display the new track condition instead and the forms for leaderboards and tracks were updated to use the new track condition field. Updated the models to remove unused validation and scopes. In addition, the track_controller was refactored to authorize the new track before saving it. This refactor simplifies the codebase and improves maintenance.