Remove weather conditions and update views
Removed 'track_weather_conditions' and 'track_times_of_day' columns from the 'tracks' table, providing a streamlined approach to track condition information. This was reflected in updates to the views for leaderboard, forms for leaderboard and tracks, track_controller and validations and scopes in models. The amendments simplify the codebase, improve maintenance and provide an updated display of the track condition.