Refactor leaderboard view display for track name
Removed wrapper class for the track name in the leaderboard view, encapsulating the text inside a div element instead. This was done to clarify the structure of the HTML element and make styling more manageable in the future. The layout and visibility of the track name remain the same, but the container element has been changed from link_to element to a div element.